Move mock data for step out of query
