+++ TeacherWorkout.Api/TeacherWorkout.Api.csproj
@@ -19,4 +19,8 @@
     <ProjectReference Include="..\TeacherWorkout.MockData\TeacherWorkout.MockData.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="GraphQL\Mock" />
+  </ItemGroup>
+
 </Project>
