+++ TeacherWorkout.Domain/Lessons/IStepRepository.cs
@@ -1,0 +1,9 @@
+using TeacherWorkout.Domain.Models;
+
+namespace TeacherWorkout.Domain.Lessons
+{
+    public interface IStepRepository
+    {
+        ILessonStep Find(string id);
+    }
+}
