+++ TeacherWorkout.Domain/Lessons/LessonStatusFindInput.cs
@@ -1,0 +1,7 @@
+namespace TeacherWorkout.Domain.Lessons
+{
+    public class StepFindInput
+    {
+        public string Id { get; set; }
+    }
+}
