+++ TeacherWorkout.Domain/Themes/GetThemes.cs
@@ -1,6 +1,5 @@
 using TeacherWorkout.Domain.Common;
 using TeacherWorkout.Domain.Models;
-using TeacherWorkout.Domain.Models.Inputs;
 
 namespace TeacherWorkout.Domain.Themes
 {
